+++ src/elements/button/tertiary-button-link/__snapshots__/tertiary-button-link.snapshot.spec.snap.js
@@ -67,6 +67,7 @@ snapshots["sbb-tertiary-button-link renders a disabled sbb-tertiary-button-link 
   aria-disabled="true"
   class="sbb-action-base sbb-tertiary-button-link"
   href="https://www.sbb.ch"
+  tabindex="-1"
 >
   <slot name="icon">
   </slot>