+++ src/elements/core/base-elements/action-base-element.ts
@@ -1,7 +1,6 @@
 import { html, LitElement, type TemplateResult } from 'lit';
 
 import { hostAttributes } from '../decorators.js';
-import { getLocalName } from '../dom.js';
 
 /**
  * Whenever an element can be disabled it has disabled property
@@ -11,6 +10,7 @@ import { getLocalName } from '../dom.js';
 type MaybeDisabled = {
   disabled?: boolean;
   formDisabled?: boolean;
+  disabledInteractive?: boolean;
 };
 
 @hostAttributes({
@@ -20,6 +20,10 @@ export abstract class SbbActionBaseElement extends LitElement {
   protected get maybeDisabled(): boolean | undefined {
     const maybeDisabled = this as MaybeDisabled;
     return maybeDisabled.disabled || maybeDisabled.formDisabled;
+  }
+
+  protected get maybeDisabledInteractive(): boolean | undefined {
+    return (this as MaybeDisabled).disabledInteractive;
   }
 
   public override connectedCallback(): void {
@@ -41,7 +45,7 @@ export abstract class SbbActionBaseElement extends LitElement {
     this.addEventListener(
       'click',
       (event) => {
-        if (this.maybeDisabled) {
+        if (this.maybeDisabled && !this.maybeDisabledInteractive) {
           event.preventDefault();
           event.stopImmediatePropagation();
         }
@@ -59,10 +63,6 @@ export abstract class SbbActionBaseElement extends LitElement {
 
   /** Default render method for button-like components. */
   protected override render(): TemplateResult {
-    return html`
-      <span class="sbb-action-base ${this.localName ?? getLocalName(this)}">
-        ${this.renderTemplate()}
-      </span>
-    `;
+    return html` <span class="sbb-action-base ${this.localName}">${this.renderTemplate()}</span> `;
   }
 }