+++ src/elements/menu/menu-button/menu-button.visual.spec.ts
@@ -37,7 +37,9 @@ describe(`sbb-menu-button`, () => {
           ?disabled=${disabled}
         >
           ${label} ${index}
-          ${slottedIcon ? html`<sbb-icon slot="icon" name="pie-small"></sbb-icon>` : nothing}
+          ${slottedIcon
+            ? html`<sbb-icon slot="icon" name="face-smiling-small"></sbb-icon>`
+            : nothing}
         </sbb-menu-button>
       `,
     )}