+++ src/elements/slider/slider.spec.ts
@@ -82,4 +82,21 @@ describe(`sbb-slider`, () => {
     await keyboardPressTimes(element, 'ArrowDown');
     expect(changeEvent.count).not.to.be.greaterThan(0);
   });
+
+  it('should not be focused when disabled', async () => {
+    expect(element.tabIndex).to.be.equal(0);
+    expect(element.getAttribute('aria-disabled')).to.be.null;
+
+    element.toggleAttribute('disabled', true);
+    await waitForLitRender(element);
+
+    expect(element.tabIndex).to.be.equal(-1);
+    expect(element.getAttribute('aria-disabled')).to.be.equal('true');
+
+    element.toggleAttribute('disabled', false);
+    await waitForLitRender(element);
+
+    expect(element.tabIndex).to.be.equal(0);
+    expect(element.getAttribute('aria-disabled')).to.be.null;
+  });
 });